requirements.txt
################################################################################
# This requirements file has been automatically generated from `Pipfile` with
# `pipenv-to-requirements`
#
#
# This has been done to maintain backward compatibility with tools and services
# that do not support `Pipfile` yet.
#
# Do NOT edit it directly, use `pipenv install [-d]` to modify `Pipfile` and
# `Pipfile.lock` and then regenerate `requirements*.txt`.
################################################################################

django==2.2.5
djangorestframework==3.12.2
pillow==8.1.0
pyjwt==2.0.0
pytz==2020.5
sqlparse==0.4.1 ; python_version >= '3.5'
